Add .size directives for most global functions.
The .size directives enable looking at the .elf file with "nm" and similar tools to judge which functions are large and might deserve optimization.
(Note that this isn't perfect, mostly because of the way the individual commands are implemented as inline code rather than functions.  so the length of "start" and "main_prompt" includes the sum of all the cmd_xxx code.)

Also, "fix" the multi-line string in cmd_H.S to prevent assembler warnings.
(Hmm.  I notice that "\n" is being used as a line terminator.  Should this be "\r\n", especially for real hardware?)

This