Add CRLF support
Some systems require outputting an explicit CR before each LF to get
output formatted correctly.  This adds support for that, and makes it
the default.

Note that only `print_newline` and `print_string` (and the
`uart_output_string` function it depends on) do this check; if you print
any other way it will output the bytes you specify.